Move the common message outside the service folder
