+++ x-pack/plugins/apm/public/components/shared/links/apm/service_link/index.tsx
@@ -19,7 +19,7 @@ import { ApmRoutes } from '../../../../routing/apm_route_config';
 import { AgentIcon } from '../../../agent_icon';
 import { PopoverTooltip } from '../../../popover_tooltip';
 import { TruncateWithTooltip } from '../../../truncate_with_tooltip';
-import { OTHER_SERVICE_NAME, MaxGroupsMessage } from './max_groups_message';
+import { OTHER_SERVICE_NAME, MaxGroupsMessage } from '../max_groups_message';
 
 const StyledLink = euiStyled(EuiLink)`${truncate('100%')};`;
