+++ x-pack/plugins/apm/public/components/shared/links/apm/transaction_detail_link/index.tsx
@@ -19,7 +19,7 @@ import { getComparisonEnabled } from '../../../time_comparison/get_comparison_en
 import { TruncateWithTooltip } from '../../../truncate_with_tooltip';
 import { APMQueryParams } from '../../url_helpers';
 import { APMLinkExtendProps, getLegacyApmHref } from '../apm_link';
-import { MaxGroupsMessage } from '../service_link/max_groups_message';
+import { MaxGroupsMessage } from '../max_groups_message';
 
 export const txGroupsDroppedBucketName = '_other';
